requirements.txt
fastapi
uvicorn
celery
redis
sentence-transformers
pydantic
pydantic-settings
python-dotenv
sqlalchemy[asyncio]
asyncpg
pgvector
alembic
